criacao do componente de visualizar e editar veterinario e correcoes no agendamento

+++ backend/veterinaria/src/main/java/br/com/softvet/veterinaria/model/repository/VeterinarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.softvet.veterinaria.model.entity.Veterinario;
 
-public interface VeterinarioRepository extends JpaRepository<Veterinario, Integer>{
+public interface VeterinarioRepository extends JpaRepository<Veterinario, Long>{
 
 }
